+++ HISTORY.rst
@@ -9,3 +9,5 @@ History
 * Add an API client to wrap the Data for SEO RestClient
 * Add ability to get credentials from environment variables
 ** DFS_API_USERNAME and DFS_API_PASSWORD
+* Add service to get locations and languages
+* Add service to get domain competitors